reworked the game loop
- just not gonna have a general update because fuck that fixed update only go fuck yourself

- moved window creation, sdl init and polling to the main thread since cocoa is super anal ab that so hopefully this'll work on the macbook now uwu

- more renderer preprocessor gating

- added smth for time dilation not absolutely sure how to do that yet but w/e

if u want smooth graphics or whatever the render thread runs fully independent in its own hot loop, respectfully tho it doesn't spin forever eating cycles and doing the OS scheduler dirty owo

and just use a tween that i'll setup in the engine so that it'll lerp over all render frames similiar to godot's tween so u don't need to use a general Update() w delta time garbage to do stuff

mainly because having the frame stuff be tightly coupled just seems better, like general updates can just cause weird issues and require heuristics to figure out if i should call update again since consuming update on the main thread can cause fixed updates to not be done at the correct interval and in the worst case scenario really muck up stuff for processes that require rigid time deltas idk like it's not that big of a deal but still bothers me

even if i delegate running scripts to another thread its not like i can invoke scripts concurrently without breaking state and theres no consistent way that feels good and isn't absurd to let devs know how state is being handled if they have a script w update and fixed update idk

because if update lags it'll kill physics, but it wont affect rendering anymore since that's fully decoupled now and we can do that because rendering doesn't impact state, it just reads state from static memory chunks that don't move and realloc and draw based off that the most current state of the scene it has owo

+++ src/peach_core/Utils/Logger.cpp
@@ -16,7 +16,7 @@
 #include <fmt/format.h>
 
 static constexpr uint32_t FLUSH_EVERY_N_LOGS = 256u;
-static constexpr uintmax_t MAX_LOG_FILE_SIZE_BYTES = 10u * 1024u * 1024u; // 10 MB
+static constexpr uint32_t MAX_LOG_FILE_SIZE_BYTES = 10u * 1024u * 1024u; // 10 MB
 
 static constexpr uint8_t FLUSH_TRACE_BIT = 1u << 0;
 static constexpr uint8_t FLUSH_DEBUG_BIT = 1u << 1;
